+++ src/components/analyser/search/Filter.tsx
@@ -1,134 +1,17 @@
-/* eslint-disable no-param-reassign */
 import React from 'react';
 import {
-  createStyles,
-  makeStyles,
-  Theme,
-  useTheme,
-} from '@material-ui/core/styles';
-import {
-  Input,
-  InputLabel,
-  MenuItem,
-  FormControl,
-  ListItemText,
-  Select,
-  Checkbox,
-  Chip,
   Button,
   ButtonGroup,
 } from '@material-ui/core';
 import { FilterList, Delete } from '@material-ui/icons';
 import { useTranslation } from 'react-i18next';
 import * as API from '../../../analyser/APIClient';
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    typography: {
-      color: '#0D1B3B',
-    },
-    formControl: {
-      margin: theme.spacing(2),
-      minWidth: 200,
-      maxWidth: 300,
-    },
-    chips: {
-      display: 'flex',
-      flexWrap: 'wrap',
-    },
-    chip: {
-      margin: 1,
-    },
-  })
-);
-
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      Width: 300,
-    },
-  },
-};
-
-function getStyles(tmp: string, list: string[], theme: Theme) {
-  return {
-    fontWeight:
-      list.indexOf(tmp) === -1
-        ? theme.typography.fontWeightRegular
-        : theme.typography.fontWeightMedium,
-  };
-}
+import { FilterBar } from './FilterBar'
 
 type Filters = {
   countries: string[];
   industries: string[];
   currency: string[];
-};
-
-export type FilterBarProps = {
-  filtersList: string[];
-  ogFiltersList: string[];
-  handleChange: (
-    event: React.ChangeEvent<{ name?: string | undefined; value: unknown }>
-  ) => void;
-  name: string;
-};
-export const FilterBar: React.FC<FilterBarProps> = ({
-  filtersList,
-  ogFiltersList,
-  handleChange,
-  name,
-}) => {
-  const classes = useStyles();
-  const theme = useTheme();
-
-  return (
-    <FormControl className={classes.formControl}>
-      <InputLabel id="id-label">{name}</InputLabel>
-      <Select
-        name={name}
-        multiple
-        value={filtersList}
-        onChange={handleChange}
-        input={<Input id="select-multiple-chip" />}
-        renderValue={(selected) => (
-          <div className={classes.chips}>
-            {(selected as string[]).map((value) => (
-              <Chip
-                key={value}
-                label={value}
-                variant="outlined"
-                color="primary"
-                size="small"
-                className={classes.chip}
-              />
-            ))}
-          </div>
-        )}
-        MenuProps={MenuProps}
-      >
-        <MenuItem disabled value="">
-          <em>{name}</em>
-        </MenuItem>
-        <MenuItem value="">
-          <em>None</em>
-        </MenuItem>
-        {ogFiltersList.map((tmp) => (
-          <MenuItem
-            key={tmp}
-            value={tmp}
-            style={getStyles(tmp, filtersList, theme)}
-          >
-            <Checkbox color="primary" checked={filtersList.indexOf(tmp) > -1} />
-            <ListItemText primary={tmp} />
-          </MenuItem>
-        ))}
-      </Select>
-    </FormControl>
-  );
 };
 
 function setOriginalFilters(stocks: API.Stock[]) {
@@ -161,12 +44,7 @@ const Filteroptions: React.FC<FilterOptionsProps> = ({ stocks }) => {
   const [filters, setFilters] = React.useState<Filters>(emptyFilters);
 
   const clearFilters = () => {
-    setFilters((prevFilters) => {
-      prevFilters.countries = [];
-      prevFilters.industries = [];
-      prevFilters.currency = [];
-      return { ...prevFilters };
-    });
+    setFilters({ countries: [], industries: [], currency: [] });
   };
 
   const handleChange = (
@@ -255,7 +133,7 @@ const Filteroptions: React.FC<FilterOptionsProps> = ({ stocks }) => {
         >
           Adapt
         </Button>
-        <Button type="button" onClick={clearFilters} startIcon={<Delete />}>
+        <Button type="button" onClick={clearFilters} startIcon={<Delete/>}>
           Clear All
         </Button>
       </ButtonGroup>
